refactor: more shipping package from `no-meteor` to `node-app`
