+++ imports/node-app/plugins/shipping/i18n/fr.json
