+++ imports/node-app/plugins/shipping/i18n/ru.json
