+++ imports/node-app/plugins/shipping/i18n/sl.json
