+++ imports/node-app/plugins/shipping/index.js
@@ -1,8 +1,8 @@
-import i18n from "./i18n";
-import mutations from "./mutations";
-import queries from "./queries";
-import resolvers from "./resolvers";
-import schemas from "./schemas";
+import i18n from "./i18n/index.js";
+import mutations from "./mutations/index.js";
+import queries from "./queries/index.js";
+import resolvers from "./resolvers/index.js";
+import schemas from "./schemas/index.js";
 
 /**
  * @summary Import and call this function to add this plugin to your API.
