+++ imports/node-app/plugins/shipping/mutations/index.js
@@ -1,0 +1,7 @@
+import selectFulfillmentOptionForGroup from "./selectFulfillmentOptionForGroup.js";
+import updateFulfillmentOptionsForGroup from "./updateFulfillmentOptionsForGroup.js";
+
+export default {
+  selectFulfillmentOptionForGroup,
+  updateFulfillmentOptionsForGroup
+};
