+++ imports/node-app/plugins/shipping/queries/getFulfillmentMethodsWithQuotes.js
@@ -1,5 +1,5 @@
 import Logger from "@reactioncommerce/logger";
-import extendCommonOrder from "../util/extendCommonOrder";
+import extendCommonOrder from "../util/extendCommonOrder.js";
 
 /**
  * @name getFulfillmentMethodsWithQuotes
