+++ imports/node-app/plugins/shipping/queries/index.js
@@ -1,0 +1,5 @@
+import getFulfillmentMethodsWithQuotes from "./getFulfillmentMethodsWithQuotes.js";
+
+export default {
+  getFulfillmentMethodsWithQuotes
+};
