+++ imports/node-app/plugins/shipping/resolvers/Mutation/index.js
@@ -1,0 +1,7 @@
+import selectFulfillmentOptionForGroup from "./selectFulfillmentOptionForGroup.js";
+import updateFulfillmentOptionsForGroup from "./updateFulfillmentOptionsForGroup.js";
+
+export default {
+  selectFulfillmentOptionForGroup,
+  updateFulfillmentOptionsForGroup
+};
