--- imports/plugins/core/shipping/server/no-meteor/mutations/index.js
@@ -1,7 +1,0 @@
-import selectFulfillmentOptionForGroup from "./selectFulfillmentOptionForGroup";
-import updateFulfillmentOptionsForGroup from "./updateFulfillmentOptionsForGroup";
-
-export default {
-  selectFulfillmentOptionForGroup,
-  updateFulfillmentOptionsForGroup
-};
