--- imports/plugins/core/shipping/server/no-meteor/queries/index.js
@@ -1,5 +1,0 @@
-import getFulfillmentMethodsWithQuotes from "./getFulfillmentMethodsWithQuotes";
-
-export default {
-  getFulfillmentMethodsWithQuotes
-};
